+++ app/models/hydra/access_controls/lease_decorator.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# OVERRIDE Hydra-access-controls 12.0.1
+# OVERRIDE Hydra-access-controls 12.1.0
 # Fix releasing leases on the day they are expired - this solves a 1 second bug around how
 # midnights are calculated, which causes day of leases to incorrectly set the permissions to private
 module Hydra
